requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
django-crispy-forms==2.1
crispy-bootstrap5==2025.6


# Database
psycopg2-binary==2.9.9
pymongo==4.6.0

# Authentication & Security
PyJWT==2.8.0
python-decouple==3.8

# API & HTTP
requests==2.31.0
httpx==0.25.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
python-dateutil==2.8.2

# Development
ipython==8.17.2
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0